backend/requirements.txt
torch==2.0.1
torchvision==0.15.2
flask==2.3.2
flask-cors==4.0.0
pillow==10.0.0
numpy==1.24.3
matplotlib==3.7.2